# Performance improvements based on benchmarks