--- a/src/Treasure.Utils.Argument/Argument.cs
+++ b/src/Treasure.Utils.Argument/Argument.cs
@@ -27,10 +27,14 @@ public static class Argument
     /// <returns>The value if not null.</returns>
     public static T NotNull<T>([NotNull] T? value, [CallerArgumentExpression(nameof(value))] string name = "") where T : class
     {
+#if NET6_0_OR_GREATER
+        ArgumentNullException.ThrowIfNull(value, name);
+#else
         if (value is null)
         {
             throw new ArgumentNullException(name);
         }
+#endif
 
         return value;
     }
@@ -43,6 +47,9 @@ public static class Argument
     /// <returns><see cref="string"/>.</returns>
     public static string NotNullOrEmpty([NotNull] string? value, [CallerArgumentExpression(nameof(value))] string name = "")
     {
+#if NET7_0_OR_GREATER
+        ArgumentException.ThrowIfNullOrEmpty(value, name);
+#else
         if (value is null)
         {
             throw new ArgumentNullException(name);
@@ -52,6 +59,7 @@ public static class Argument
         {
             throw new ArgumentException("The value cannot be null or empty.", name);
         }
+#endif
 
         return value;
     }