--- a/test/Directory.Packages.props
+++ b/test/Directory.Packages.props
@@ -3,6 +3,7 @@
   <Import Project="$([MSBuild]::GetPathOfFileAbove(Directory.Packages.props, $(MSBuildThisFileDirectory)..))" />
 
   <ItemGroup>
+    <PackageVersion Include="BenchmarkDotNet"                                   Version="0.13.8" />
     <PackageVersion Include="coverlet.collector"                                Version="6.0.0" />
     <PackageVersion Include="Microsoft.NET.Test.Sdk"                            Version="17.7.2" />
     <PackageVersion Include="xunit"                                             Version="2.5.1" />