--- a/test/Treasure.Utils.Argument.Benchmarks/Treasure.Utils.Argument.Benchmarks.csproj
+++ b/test/Treasure.Utils.Argument.Benchmarks/Treasure.Utils.Argument.Benchmarks.csproj
@@ -1,0 +1,21 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFrameworks>net7.0;net6.0</TargetFrameworks>
+    <RootNamespace>Treasure.Utils.Benchmarks</RootNamespace>
+    <IsTestProject>false</IsTestProject>
+
+    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
+    <ServerGarbageCollection>true</ServerGarbageCollection>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="BenchmarkDotNet" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\src\Treasure.Utils.Argument\Treasure.Utils.Argument.csproj" />
+  </ItemGroup>
+
+</Project>