While performing the .NET 8 upgrade, I decided to put together some benchmarks to help optimize the library a bit more. They seemed good enough to extract out to the current TFMs. After adding the benchmarks, I performed the following optimizations and measured:

- Per TFM optimizations to use new platform specific helpers.
- Add suggestions to inline the methods.

The original measurements based on the changes before the optimizations in this PR:

| Method              | Runtime  | Mean      | Ratio | Code Size |
|-------------------- |--------- |----------:|------:|----------:|
| NotNull             | .NET 6.0 | 0.0047 ns |     ? |      74 B |
| NotNull             | .NET 7.0 | 0.0281 ns |     ? |      75 B |
|                     |          |           |       |           |
| NotNullOrEmpty      | .NET 6.0 | 1.2290 ns |  1.00 |     159 B |
| NotNullOrEmpty      | .NET 7.0 | 1.2385 ns |  1.01 |     162 B |
|                     |          |           |       |           |
| NotNullOrWhiteSpace | .NET 6.0 | 2.7396 ns |  1.00 |     168 B |
| NotNullOrWhiteSpace | .NET 7.0 | 2.9706 ns |  1.08 |     172 B |

After performing the TFM optimizations:

| Method              | Runtime  | Mean      | Ratio | Code Size |
|-------------------- |--------- |----------:|------:|----------:|
| NotNull             | .NET 6.0 | 0.0121 ns |     ? |      43 B |
| NotNull             | .NET 7.0 | 0.0001 ns |     ? |      44 B |
|                     |          |           |       |           |
| NotNullOrEmpty      | .NET 6.0 | 1.2099 ns | 1.000 |     159 B |
| NotNullOrEmpty      | .NET 7.0 | 0.2334 ns | 0.193 |      58 B |
|                     |          |           |       |           |
| NotNullOrWhiteSpace | .NET 6.0 | 2.7246 ns |  1.00 |     168 B |
| NotNullOrWhiteSpace | .NET 7.0 | 2.9730 ns |  1.09 |     172 B |

This enabled .NET 7 to take advantage of some new platform APIs that now implement the same behavior to get a good speed boost on `NotNull` and `NotNullOrEmpty`.  The code complexity in `NotNullOrWhiteSpace` likely deterred inlining.

And then after adding the inline suggestions (includes TFM optimizations):

| Method              | Runtime  | Mean      | Ratio | Code Size |
|-------------------- |--------- |----------:|------:|----------:|
| NotNull             | .NET 6.0 | 0.0022 ns |     ? |      43 B |
| NotNull             | .NET 7.0 | 0.0000 ns |     ? |      44 B |
|                     |          |           |       |           |
| NotNullOrEmpty      | .NET 6.0 | 1.2180 ns | 1.000 |     159 B |
| NotNullOrEmpty      | .NET 7.0 | 0.2377 ns | 0.195 |      58 B |
|                     |          |           |       |           |
| NotNullOrWhiteSpace | .NET 6.0 | 2.7537 ns |  1.00 |     168 B |
| NotNullOrWhiteSpace | .NET 7.0 | 1.7353 ns |  0.63 |     285 B |

.NET 6 doesn't get a boost from the inlining suggestions, but .NET 7 did get a bit of a boost in the `NotNullOrWhiteSpace` method.

.NET 8 includes an implementation for `NotNullOrWhiteSpace`. While we don't get a speed boost, we do get a code size improvement.